Updated Location Paths
Added (Authed): Character.Location
Added (Authed): Character.Location.IsOnline
Modified: Character.GetLocation moved to Character.Location.GetLocation
Modified: Character.GetCurrentShip moved to Character.Location.GetCurrentShip
Modified: Character to include Character.Location class
